+++ resources/views/projets.blade.php
@@ -23,7 +23,9 @@
 						<div class="card-body">
 							<div class="titre-ajout mb-2">
 								<h4 class="header-title mt-0 mb-1">Liste des projets</h4>
-								<a class="ml-auto btn btn-primary" href="/projets/create">Ajouter</a>
+								@if (auth()->user()->role != "Visiteur")
+									<a class="ml-auto btn btn-primary" href="/projets/create">Ajouter</a>
+								@endif
 							</div>
 
 							<div class="table-responsive">
@@ -52,14 +54,18 @@
 												<td>{{$projet->libelle_projet}}</td>
 												<td>{{$projet->debut_projet}}</td>
 												<td>{{$projet->fin_projet}}</td>
-												<td>
-													<a href="/projets/{{$projet->id}}/edit" class="btn btn-primary btn-sm">
-														<i class="bi bi-pencil-square"></i>
-													</a>
-													<a href="/projets/{{$projet->id}}/delete" class="btn btn-danger btn-sm suppression">
-														<i class="bi bi-trash"></i>
-													</a>
-												</td>
+												@if (auth()->user()->role == "Administrateur")
+													<td>
+														<a href="/projets/{{$projet->id}}/edit" class="btn btn-primary btn-sm">
+															<i class="bi bi-pencil-square"></i>
+														</a>
+														<a href="/projets/{{$projet->id}}/delete" class="btn btn-danger btn-sm suppression">
+															<i class="bi bi-trash"></i>
+														</a>
+													</td>
+												@else
+													<td></td>
+												@endif
 											</tr>
 										@endforeach
 									</tbody>
